+++ src/app.js
@@ -5,8 +5,8 @@ import { responseHandler } from "./middlewares/responseHandler.js";
 import { errorHandler } from "./middlewares/errorHandler.js";
 import { setupSwagger } from "./config/swagger.js";
 import authRouter from "./routes/auth.routes.js";
-import searchRouter from "./routes/searchRoutes.js";
-import cafeRouter from "./routes/cafeRoutes.js";
+import searchRouter from "./routes/search.routes.js";
+import cafeRouter from "./routes/cafe.routes.js";
 import passport from "./config/passport.js";
 import userRouter from "./routes/user.routes.js";
 const app = express();
