+++ src/controllers/cafe.controller.js
@@ -5,7 +5,7 @@ import {
   cafeReviewService,
   cafeCouponService,
   cafeBookmarkService,
-} from "../services/cafeService.js";
+} from "../services/cafe.service.js";
 
 //북마크 여부도 체크해야함
 export const getCafe = async (req, res, next) => {
